backend/requirements-dev.txt
coverage==4.5.3
Datetime==4.3
docopts==0.6.1
pytest==5.0.1
pytest-cov==2.8.1
pytest-mock==1.11.1
python-dateutil==2.8.0
pytz==2019.3
requests==2.22.0
requests_mock==1.7.0
uuid==1.30
vcrpy==2.1.0
Flask==1.1.1
flask-marshmallow==0.10.1
redis==3.3.10
gevent==1.4.0
gunicorn==19.9.0
